requirements/base.txt
django-extra-views==0.7.1
django-filter==0.12.0
django-widget-tweaks==1.4.1
django-betterforms==1.1.4